% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_4_plot_gen.R
\name{cluster_4_plot_gen}
\alias{cluster_4_plot_gen}
\title{Generates 4 consecutive histograms (in a row) of cluster count distribution from cutoff, given 4 cutoffs.}
\usage{
cluster_4_plot_gen(cutoff1, cutoff2, cutoff3, cutoff4)
}
\arguments{
\item{cutoff1}{a hierarchical clustering cutoff, from 0 to 1, same as cluster_list_gen}

\item{cutoff4}{}
}
\value{
4 histograms of cluster count distribution, using cluster_plot_gen
}
\description{
Generates 4 consecutive histograms (in a row) of cluster count distribution from cutoff, given 4 cutoffs.
}
