% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_count_gen.R
\name{cluster_count_gen}
\alias{cluster_count_gen}
\title{Counts sequences in each assigned cluster}
\usage{
cluster_count_gen(list)
}
\arguments{
\item{list}{a list (in tibble format) generated by cluster_list_gen}
}
\value{
Generates a cluster count distribution table
}
\description{
Counts sequences in each assigned cluster
}
