% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot_gen.R
\name{cluster_plot_gen}
\alias{cluster_plot_gen}
\title{Generates a histogram of cluster count distribution from cutoff}
\usage{
cluster_plot_gen(cutoff)
}
\arguments{
\item{cutoff}{a hierarchical clustering cutoff, from 0 to 1, same as cluster_list_gen}
}
\value{
a histogram of cluster count distribution
}
\description{
Generates a histogram of cluster count distribution from cutoff
}
