% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Generates a heatmap of mash distances, reordered by UMPGA hierarchical clustering. Clusters are annotated accordingly by colors.}
\usage{
draw_heatmap(cutoff)
}
\arguments{
\item{cutoff}{a hierarchical clustering cutoff, from 0 to 1.}
}
\value{
a heatmap. Currently only supports annotating 50 clusters or less. Adjust threshold accordingly.
}
\description{
Generates a heatmap of mash distances, reordered by UMPGA hierarchical clustering. Clusters are annotated accordingly by colors.
}
