% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Read in a mash pairwise distance file, parses it and save in variable "mash_matrix" in the global environment.}
\usage{
input(filepath)
}
\arguments{
\item{filepath}{IN QUOTES: Relative path to the mash pairwise distance output file, should be a 5-column tsv without column names.}
}
\value{
parsed distance matrix saved in variable "mash_matrix" in the global environment. Do not have other variables with the same name in .GlobalEnv.
}
\description{
Read in a mash pairwise distance file, parses it and save in variable "mash_matrix" in the global environment.
}
