% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_num_list.R
\name{total_num_list}
\alias{total_num_list}
\title{Generates a chart of possible total number of cluster and the minimum cutoff required to reach that number. No input is needed.}
\usage{
total_num_list()
}
\value{
Generates a chart of possible total number of cluster and the minimum cutoff used to reach that number.
}
\description{
Generates a chart of possible total number of cluster and the minimum cutoff required to reach that number. No input is needed.
}
