% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_num_plot.R
\name{total_num_plot}
\alias{total_num_plot}
\title{Generates a geom_step plot showing distribution of total cluster counts (y axis) by clustering cutoff (x axis).}
\usage{
total_num_plot(min = 0.01, max = 1, log = TRUE)
}
\arguments{
\item{min}{The minimum cutoff shown in this plot. Defaults to 0.01.}

\item{max}{The maximum cutoff shown in this plot. Defaults to 1.}

\item{log}{A logical value (TRUE/FALSE) to indicate if the logarithm scale should be used on the cutoff (x) axis. Defaults to FALSE}
}
\value{
A step plot showing how cluster count decreases over raising the cutoff.
}
\description{
Generates a geom_step plot showing distribution of total cluster counts (y axis) by clustering cutoff (x axis).
}
