% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_list_gen.R
\name{cluster_list_gen}
\alias{cluster_list_gen}
\title{Generates a UPGMA hierarchical clustered table of Mash sequences based on given threshold from 0 to 1.}
\usage{
cluster_list_gen(cutoff)
}
\arguments{
\item{cutoff}{The mash distance cutoff for hierarchical clustering, from 0 to 1. cutoff=0 means sequences will be divided when their mash distance is above 0. cutoff >=1 means all sequences will form one cluster.}
}
\value{
A clustering list (tibble format). The first column is the mash sequences in original order, the second column is the assigned cluster bases on given cutoff.
}
\description{
Generates a UPGMA hierarchical clustered table of Mash sequences based on given threshold from 0 to 1.
}
